requirements.txt
Flask==2.3.2 # Or your installed Flask version
requests==2.31.0 # Or your installed requests version
python-dotenv==1.0.0 # Or your installed python-dotenv version